+++ packages/nadle/src/builtin-tasks/npx-task.ts
@@ -1,5 +1,5 @@
-import { runCommand } from "./run-command.js";
-import { MaybeArray } from "../core/index.js";
+import { type MaybeArray } from "../core/index.js";
+import { runCommand, normalizeArgs } from "./run-command.js";
 import { defineTask } from "../core/registration/define-task.js";
 
 /**
@@ -19,12 +19,10 @@ export interface NpxTaskOptions {
  */
 export const NpxTask = defineTask<NpxTaskOptions>({
 	run: async ({ options, context }) => {
-		const args = options.args == null ? [] : MaybeArray.toArray(options.args);
-
 		await runCommand(context, {
 			command: "npx",
-			args: [options.command, ...args],
 			doneMessage: `npx command completed successfully.`,
+			args: [options.command, ...normalizeArgs(options.args)],
 			startMessage: (finalArgs) => `Running npx command: npx ${finalArgs.join(" ")}`
 		});
 	}